LayoutTests/fast/dom/icon-url-property-expected.txt

main frame - didChangeIcons
Original iconURL is: http://test.com/oldfavicon.ico
Setting new icon URL to: http://test.com/newfavion.ico
New iconURL is: http://test.com/newfavion.ico
